TIP-842: Move family variant normalizer to Akeneo\Pim\Structure
