+++ src/Akeneo/Tool/Bundle/ApiBundle/Resources/config/normalizers.yml
@@ -3,7 +3,6 @@ parameters:
     pim_api.normalizer.exception.violation.class: Akeneo\Tool\Component\Api\Normalizer\Exception\ViolationNormalizer
     pim_api.normalizer.collection.class: Akeneo\Tool\Component\Api\Normalizer\CollectionNormalizer
     pim_api.normalizer.product.class: Akeneo\Tool\Component\Api\Normalizer\ProductNormalizer
-    pim_api.normalizer.family_variant.class: Akeneo\Tool\Component\Api\Normalizer\FamilyVariantNormalizer
     pim_api.normalizer.category.class: Akeneo\Tool\Component\Api\Normalizer\CategoryNormalizer
     pim_api.normalizer.attribute_group.class: Akeneo\Tool\Component\Api\Normalizer\AttributeGroupNormalizer
     pim_api.normalizer.file.class: Akeneo\Tool\Component\Api\Normalizer\FileNormalizer
@@ -44,13 +43,6 @@ services:
             - '@pim_catalog.normalizer.standard.product'
             - '@pim_api.repository.attribute'
             - '@router'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
-
-    pim_api.normalizer.family_variant:
-        class: '%pim_api.normalizer.family_variant.class%'
-        arguments:
-            - '@pim_catalog.normalizer.standard.family_variant'
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
