+++ src/Pim/Bundle/CatalogBundle/Resources/config/serializers_standard.yml
@@ -12,7 +12,6 @@ parameters:
     pim_catalog.normalizer.standard.association_type.class: Pim\Component\Catalog\Normalizer\Standard\AssociationTypeNormalizer
     pim_catalog.normalizer.standard.attribute_group.class: Pim\Component\Catalog\Normalizer\Standard\AttributeGroupNormalizer
     pim_catalog.normalizer.standard.category.class: Pim\Component\Catalog\Normalizer\Standard\CategoryNormalizer
-    pim_catalog.normalizer.standard.family_variant.class: Pim\Component\Catalog\Normalizer\Standard\FamilyVariantNormalizer
     pim_catalog.normalizer.standard.group.class: Pim\Component\Catalog\Normalizer\Standard\GroupNormalizer
     pim_catalog.normalizer.standard.group_type.class: Pim\Component\Catalog\Normalizer\Standard\GroupTypeNormalizer
     pim_catalog.normalizer.standard.translation.class: Pim\Component\Catalog\Normalizer\Standard\TranslationNormalizer
@@ -99,13 +98,6 @@ services:
 
     pim_catalog.normalizer.standard.category:
         class: '%pim_catalog.normalizer.standard.category.class%'
-        arguments:
-            - '@pim_catalog.normalizer.standard.translation'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
-
-    pim_catalog.normalizer.standard.family_variant:
-        class: '%pim_catalog.normalizer.standard.family_variant.class%'
         arguments:
             - '@pim_catalog.normalizer.standard.translation'
         tags:
